# Conflicts:
#	.gitignore
#	Project/.project
#	Project/.settings/org.eclipse.wst.common.component
#	Project/.settings/org.eclipse.wst.common.project.facet.core.xml
#	Project/src/main/java/com/project/controller/jmj/myPageController.java
#	Project/src/main/java/com/project/controller/kjs/ShoppingCartController.java
#	Project/src/main/java/com/project/controller/kkb/admin/member/MemberAdminController.java
#	Project/src/main/java/com/project/controller/ksh/OrderController.java
#	Project/src/main/java/com/project/dao/kkb/admin/AdminMemberDAO.java
#	Project/src/main/java/com/project/dao/kkb/admin/AdminMemberDAOImpl.java
#	Project/src/main/java/com/project/dao/member/MemberDAO.java
#	Project/src/main/java/com/project/dao/member/MemberDAOImpl.java
#	Project/src/main/java/com/project/service/kkb/admin/AdminMemberService.java
#	Project/src/main/java/com/project/service/kkb/admin/AdminMemberServiceImpl.java
#	Project/src/main/java/com/project/service/member/MemberService.java
#	Project/src/main/java/com/project/service/member/MemberServiceImpl.java
#	Project/src/main/java/com/project/vodto/Member.java
#	Project/src/main/java/com/project/vodto/ShippingAddress.java
#	Project/src/main/java/com/project/vodto/kjy/Memberkjy.java
#	Project/src/main/java/com/project/vodto/kkb/MemberResponse.java
#	Project/src/main/java/com/project/vodto/kkb/SearchMemberRequest.java
#	Project/src/main/resources/mappers/LoginMapper.xml
#	Project/src/main/resources/mappers/adminMemberMapper.xml
#	Project/src/main/resources/mappers/memberMapper.xml
#	Project/src/main/resources/mybatis-config.xml
#	Project/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
#	Project/src/main/webapp/WEB-INF/views/header.jsp
#	Project/src/main/webapp/WEB-INF/views/user/myPage.jsp
#	Project/src/main/webapp/WEB-INF/views/user/withdrawal.jsp
#	Project/target/classes/mappers/listMapper.xml
#	Project/target/classes/mappers/memberMapper.xml
#	Project/target/classes/mappers/shoppingCartMapper.xml
#	Project/target/classes/mybatis-config.xml
#	Project/target/m2e-wtp/web-resources/META-INF/maven/com.project/controller/pom.xml

+++ Project/src/main/java/com/project/controller/kjs/ShoppingCartController.java
@@ -1,6 +1,5 @@
 package com.project.controller.kjs;
 
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.sql.SQLException;
 import java.util.HashMap;
@@ -17,13 +16,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.util.WebUtils;
@@ -78,7 +74,8 @@ public class ShoppingCartController {
 		try {
 			if (session.getAttribute("loginMember") != null) {
 				// 로그인을 했으면
-				list = scService.getShoppingCart(((Member)session.getAttribute("loginMember")).getMemberId(), true);
+				Member member = (Member) session.getAttribute("loginMember");
+				list = scService.getShoppingCart(member.getMemberId(), true);
 				List<Product> items = (List<Product>)list.get("items");
 				result = output(list, items);
 			} else {
@@ -141,26 +138,34 @@ public class ShoppingCartController {
 		ResponseEntity<Map<String, Object>> result = null;
 		Map<String, Object> map = new HashMap<String, Object>();
 		// 로그인 여부 확인
-		// 로그인을 했으면
+		HttpSession session = request.getSession();
 		
-		// 비회원이면
-		Cookie cookie = WebUtils.getCookie(request, "nom");
-		System.out.println("id : " + cookie.getValue() + ", productId : " + productId);
-		if (cookie != null) {
-			try {				
-				// 하나 이상의 row가 삭제되었다면
-				if (scService.deleteItem(cookie.getValue(), false, productId)) {
+		try {				
+			if (session.getAttribute("loginMember") != null) {
+				// 로그인을 했으면
+				if (scService.deleteItem(((Member)session.getAttribute("loginMember")).getMemberId(), true, productId)) {
 					map.put("status", "success");
+					result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
 				}
-				else {
-					map.put("status", "fail");
+			} else {
+				// 비회원이면
+				Cookie cookie = WebUtils.getCookie(request, "nom");
+				System.out.println("id : " + cookie.getValue() + ", productId : " + productId);
+				if (cookie != null) {
+					// 하나 이상의 row가 삭제되었다면
+					if (scService.deleteItem(cookie.getValue(), false, productId)) {
+						map.put("status", "success");
+					}
+					else {
+						map.put("status", "fail");
+					}
+					result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
 				}
-				result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
-			} catch (SQLException | NamingException e) {
-				map.put("status", "fail");
-				result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.BAD_REQUEST);
-				e.printStackTrace();
 			}
+		} catch (SQLException | NamingException e) {
+			map.put("status", "fail");
+			result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.BAD_REQUEST);
+			e.printStackTrace();
 		}
 		
 		System.out.println("======= 장바구니 컨트롤러단 끝 =======");
@@ -201,26 +206,38 @@ public class ShoppingCartController {
 	}
 	
 	// item 추가
-	@RequestMapping(value="{productId}", method=RequestMethod.POST)
-	public ResponseEntity<Map<String, Object>> addProduct(HttpServletRequest request, @PathVariable("productId") String productId) {
+	@RequestMapping("insert")
+	public ResponseEntity<Map<String, Object>> addProduct(HttpServletRequest request, @RequestParam("product_id") String productId) {
 		System.out.println("======= 장바구니 컨트롤러 - 아이템 추가 =======");
 		ResponseEntity<Map<String, Object>> result = null;
 		Map<String,Object> map = new HashMap<String, Object>();
 		// 로그인 여부 확인
-		// 로그인 했을 시
-		
-		// 비회원일 시
-		Cookie cookie = WebUtils.getCookie(request, "nom");
-		if (cookie != null) {
-			try {
-				scService.insertItem(cookie.getValue(), false, productId);
-				map.put("status", "success");
-				result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
-			} catch (SQLException | NamingException e) {
-				e.printStackTrace();
-				map.put("status", "fail");
-				result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.BAD_REQUEST);
+		HttpSession session = request.getSession();
+		try {
+			if (session.getAttribute("loginMember") != null) {
+				// 로그인 했을 시
+				if (scService.insertItem(((Member)session.getAttribute("loginMember")).getMemberId(), true, productId)) {
+					map.put("status", "success");
+					result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
+				} else {
+					// 뭔가의 이유로 실패 시
+				}
+			} else {
+				// 비회원일 시
+				Cookie cookie = WebUtils.getCookie(request, "nom");
+				if (cookie != null) {
+					if (scService.insertItem(cookie.getValue(), false, productId)) {
+						map.put("status", "success");
+						result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
+					} else {
+						// 무언가의 이유로 실패 시
+					}
+				}
 			}
+		} catch (SQLException | NamingException e) {
+			e.printStackTrace();
+			map.put("status", "fail");
+			result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.BAD_REQUEST);
 		}
 		
 //		String history = request.getHeader("referer");
@@ -232,9 +249,12 @@ public class ShoppingCartController {
 		return result;
 	}
 	
-	@RequestMapping("/insert")
-	public ResponseEntity<String> test(@RequestParam("product_id") String id){
-		System.out.println(id + "장바구니 들어옴");
-		return ResponseEntity.ok("Success");
+	// 결제페이지에 데이터 넘기기
+	public void orderPage(HttpServletRequest request) {
+		System.out.println("======= 장바구니 컨트롤러 - 결제페이지 이동 =======");
+		
+		
+		
+		System.out.println("======= 장바구니 컨트롤러단 끝 =======");
 	}
 }
